+++ internal/bootstrap/kafka.go
@@ -44,6 +44,14 @@ type groupHeatReader interface {
 }
 
 func RegisterKafkaHandlersWithRepositories(hub kafkaWSEventSender, repos *appComposition.Repositories) error {
+	return registerCoreKafkaHandlers(hub, repos, true)
+}
+
+func RegisterCoreKafkaHandlersWithRepositories(hub kafkaWSEventSender, repos *appComposition.Repositories) error {
+	return registerCoreKafkaHandlers(hub, repos, false)
+}
+
+func registerCoreKafkaHandlers(hub kafkaWSEventSender, repos *appComposition.Repositories, includeMessagePersistence bool) error {
 	if platformKafka.Subscriber == nil {
 		return nil
 	}
@@ -68,8 +76,9 @@ func RegisterKafkaHandlersWithRepositories(hub kafkaWSEventSender, repos *appCom
 	}
 	hotGroupNotifier := newHotGroupNotifyAggregator(hub, hotGroupNotifyWindow)
 
-	platformKafka.Subscriber.Register("message.direct.send_requested", persistMessageHandler(messaging.Messages, "direct"))
-	platformKafka.Subscriber.Register("message.group.send_requested", persistMessageHandler(messaging.Messages, "group"))
+	if includeMessagePersistence {
+		RegisterMessageKafkaHandlers(messaging.Messages)
+	}
 	platformKafka.Subscriber.Register("message.direct.created", updateConversationHandler(messaging.Conversations, false))
 	platformKafka.Subscriber.Register("message.group.created", updateConversationHandler(messaging.Conversations, true))
 	if hub != nil {
@@ -132,6 +141,14 @@ func RegisterKafkaHandlersWithRepositories(hub kafkaWSEventSender, repos *appCom
 	platformKafka.Subscriber.Register("contact.friend.deleted", logKafkaEventHandler("contact.friend.deleted"))
 
 	return nil
+}
+
+func RegisterMessageKafkaHandlers(persister kafkaMessagePersister) {
+	if platformKafka.Subscriber == nil || persister == nil {
+		return
+	}
+	platformKafka.Subscriber.Register("message.direct.send_requested", persistMessageHandler(persister, "direct"))
+	platformKafka.Subscriber.Register("message.group.send_requested", persistMessageHandler(persister, "group"))
 }
 
 func logKafkaEventHandler(topic string) platformKafka.Handler {